Execute direct commands to the AxiDraw plotter

These small utility functions are not concerned with plotting, but rather lets you query information about the plotter or execute small commands useful for setup.

axi_version(options = axi_options())

axi_sysinfo(options = axi_options())

axi_toggle_pen(options = axi_options())

axi_align(options = axi_options())

axi_off()

axi_pen_test(tip_size, line_overlap, options = axi_options())

axi_align_test(pen_a, pen_b)

axi_pen_align(x_offset = 0, y_offset = 0)

Arguments

options

An axi_options object. See the documentation for axi_options() for all the settings. Many are not relevant for these function.

tip_size

The size of the tip (i.e. the width of the line it draws) of the initial pen

line_overlap

The overlap between adjacent pen strokes when filling out shapes and drawing thick lines, in mm. Setting this to a negative amount will cause gaps between the lines.

pen_a, pen_b

The two pen specifications that should be tested for alignment with each other

x_offset, y_offset

The offsets in mm to test

Details

The different functions have various, unrelated, uses:

axi_version() returns the version of the AxiDraw software

axi_sysinfo() returns the EBB firmware version and system information

axi_toggle_pen() toggles the position of the pen (raises it if lowered and vice versa)

axi_align() raises the pen and turns off the stepper motors so the carriage can be moved freely.

axi_pen_test() will test the look of filled areas with respect to a specific pen setting

axi_pen_align() allows you to test different offset values between two pens to determine the correct setting.