Render an SVG file with the AxiDraw penplotter

This function takes an SVG representation of a graphic and sends it to the AxiDraw penplotter for drawing. This function uses the build-in svg drawing capabilities of the pyaxidraw library so rendering an svg with grDevices::svg() and then using axi_svg to draw it will likely produce a different result than rendering directly to axi_dev().

axi_svg(
  file,
  capture = FALSE,
  options = axi_options(),
  layer = NULL,
  copies = 1L,
  page_delay = 15,
  auto_rotate = TRUE,
  rendering = "all",
  reorder = "none",
  summary = FALSE,
  text
)

Arguments

file

An svg file

capture

Logical. Should the drawing instructions be captured and returned as a new svg.

options

An axi_options object. See the documentation for axi_options() for all the settings.

layer

The layer to plot when mode = 'layers'. See the details section

copies

The number of copies to produce

page_delay

The delay in seconds between each copy

auto_rotate

Automatically rotate the image to maximize paper usage

rendering

The type of preview to generate when preview = TRUE. Either 'none' (nothing rendered), 'down' (render when pen is down), 'up' (render when pen is up), or 'all' (render everything)

reorder

The type of automatical line reordering to perform. Either 'none' (no reordering), 'objects' (reorder within objects), 'groups' ( reorder within objects, then reorder groups), or 'full' (reorder everything together).

summary

Should time spend and distance travel be reported in the end

text

An svg as a character vector.

Value

If capture = TRUE a new svg showing the pen plotter movement that will be used for drawing.

Examples

if (FALSE) { # \dontrun{
logo <- system.file('example.svg', package = 'fawkes')
axi_svg(logo)
} # }