A mock of a manual axidraw connection

This function creates a mock of the manual axidraw connection with the same methods. The mock will collect all the instructions and provides a plot() method for inspecting the look of the output. As such it can be used to debug or test code without being connected to an AxiDraw.

axi_ghost(units = "cm", paper = "A4")

Arguments

units
Either 'cm' or 'in' giving the unit the instructions should be interpreted in.

paper
The size of the paper to plot on. Is only relevant for the plot method.

Value

An object with the methods given in the Instructions section.

Instructions

The following methods are available to the returned object:

connect(): Start a connection to the plotter.

disconnect(): End a connection to the plotter.

go_to(): Move the plotter head to the absolute (x, y) position.

move_to(): Move the plotter head to the absolute (x, y) position, raising the head if not already up.

line_to(): Move the plotter head to the absolute (x, y) position, lowering the head if not already down.

go_rel(): Move the plotter head with the relative (x, y) amount.

move_rel(): Move the plotter head with the relative (x, y) amount, raising the head if not already up.

line_rel(): Move the plotter head with the relative (x, y) amount, lowering the head if not already down.

pen_up(): Raise the pen head.

pen_down(): Lower the pen head.