Setting options for AxiDraw

This function encapsulates defining general settings for the AxiDraw. Input will be checked against valid values.

axi_options(
  speed_down = 25,
  speed_up = 75,
  acceleration = 75,
  position_down = 40,
  position_up = 60,
  rate_lower = 50,
  rate_raise = 75,
  delay_down = 0,
  delay_up = 0,
  const_speed = FALSE,
  model = NULL,
  port = NULL,
  ignore_port = FALSE
)

Arguments

speed_down
The maximum movement speed while the pen is lowered as a percentage of the maximum cariage speed. Values must lie between 1 and 110. Unless const_speed = TRUE the carriage uses a smooth accelaration meaning that the maximum speed may only be reached during long straight draws. There is little gain in plotting time when increasing this value, but it may adversely affect quality.

speed_up
As speed_down but for travel when the pen is raised. Movement when the pen is raised will always use acceleration and increasing this value can have a large effect and plotting time in cases where the pen jumps around a lot.

acceleration
Gives the acceleration and decelaretion speed as a percentage of the maximum rate. Values must lie between 1 and 100. Changing this value will not affect the maximum speed, but how fast it will reach it.

position_down
The height of the pen when lowered as a percentage of vertical travel. Value must lie between 1 and 100.

position_up
As position_down but setting the height of the pen when raised.

rate_lower
The rate of vertical movement when the pen is lowered given as a percentage of maximum vertical travel speed. Value must lie between 1 and 100.

rate_raise
As rate_lower but for specifying relative speed when raising the pen.

delay_down
A delay in ms before commencing movement after the pen has been lowered. Value must lie between -500 and 500. A delay is automatically applied based on the position and rate, but this can be modified with this setting.

delay_up
As delay_down but for commencing movement after the pen has been raised.

const_speed
A boolean given whether the carriage should travel with maximum speed while the pen is down, or use the smooth acceleration.

model
The AxiDraw model to use. This will affect the travel limits imposed on the carriage. Valid values are 'A4', 'A3', or 'XL'. If NULL the value will be taken from the 'fawkes.model' option or 'A4' if missing.

port
The port connection to use. If NULL it will be resolved automatically.

ignore_port
If set to TRUE it will always communicate with the first device located even if it is not the one given by port

Value

An axi_options object